# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_internal_domain_model.Project:
    properties:
      categories:
        items:
          $ref: '#/definitions/backend_internal_domain_model.ProjectCategory'
        type: array
      categoryColors:
        additionalProperties:
          type: string
        type: object
      categoryNames:
        items:
          type: string
        type: array
      colorBadge:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      documentCount:
        type: integer
      id:
        type: string
      isStarred:
        type: boolean
      logoUrl:
        type: string
      name:
        type: string
      role:
        description: User's role in this project (if direct member)
        type: string
      starredAt:
        type: string
      updatedAt:
        type: string
      visibility:
        type: string
      workspaceId:
        type: string
    type: object
  backend_internal_domain_model.ProjectCategory:
    properties:
      colorId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      projectId:
        type: string
      sortOrder:
        type: integer
    type: object
...
  backend_internal_domain_model.UserProfile:
    properties:
      accountNo:
        type: string
      avatarUrl:
        type: string
      bio:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      phoneNumber:
        type: string
      updatedAt:
        type: string
    type: object
...
  backend_internal_domain_model.UserSummary:
    properties:
      avatarUrl:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
    type: object
  backend_internal_domain_model.Workspace:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      logoUrl:
        type: string
      name:
        type: string
      plan:
        type: string
      role:
        description: Current user's role in this workspace
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  backend_internal_presentation_project_presenter.CreateProjectRequest:
    properties:
      categories:
        items:
          type: string
        type: array
      colorBadge:
        type: string
      description:
        type: string
      logoUrl:
        type: string
      name:
        type: string
      visibility:
        type: string
    type: object
...
  backend_internal_presentation_response.Envelope:
    properties:
      data: {}
    type: object
  backend_internal_presentation_response.ErrorEnvelope:
    properties:
      title:
        type: string
    type: object
  backend_internal_presentation_user_presenter.UpdateProfileRequest:
    properties:
      avatarUrl:
        type: string
      bio:
        type: string
      fullName:
        type: string
      phoneNumber:
        type: string
    type: object
...
  backend_internal_presentation_workspace_presenter.AddMemberRequest:
    properties:
      email:
        type: string
      role:
        type: string
    required:
    - email
    type: object
...
  backend_internal_presentation_workspace_presenter.UpdateWorkspaceRequest:
    properties:
      logoUrl:
        type: string
      name:
        type: string
      plan:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for Dokudocs backend service.
  title: Dokudocs API
  version: "1.0"
paths:
...
  /projects:
    get:
      description: Get all projects accessible to the user within the active workspace
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_domain_model.Project'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Create a new project within the active workspace
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Project creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_project_presenter.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_model.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - Project
...
  /projects/{id}/star:
    post:
      description: Toggle starred status for a project
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Project ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Toggle project star
      tags:
      - Project
...
  /user/profile:
    get:
      description: Retrieve current authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_model.UserProfile'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update profile details of authenticated user
      parameters:
      - description: Update profile payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_user_presenter.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_model.UserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
  /user/search:
    get:
      description: Search users by email or name query
      parameters:
      - description: Search query string
        in: query
        name: q
        type: string
      - default: 10
        description: Max results count
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_domain_model.UserSummary'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - User
...
  /workspaces/{id}:
    delete:
      description: Permanently delete a workspace (owner only)
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete workspace
      tags:
      - Workspace
    get:
      description: Get details of a workspace by ID
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_model.Workspace'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get workspace
      tags:
      - Workspace
    put:
      consumes:
      - application/json
      description: Update workspace details (name, plan, logo)
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Workspace update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_workspace_presenter.UpdateWorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_model.Workspace'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update workspace
      tags:
      - Workspace
  /workspaces/{id}/invites:
    post:
      consumes:
      - application/json
      description: Add a member by email to the workspace
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Add member payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_workspace_presenter.AddMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add workspace member
      tags:
      - Workspace
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"